Time computing is fixed. Time is everywhere set as float type, and in seconds (by default). Function 'getElapsedTime' returns elapsed time as float type.

+++ include/configurations/RGBMTStarConfig.h
@@ -7,6 +7,6 @@ class RGBMTStarConfig
 public:
     static unsigned long MAX_NUM_ITER;          // Maximal number of algorithm iterations
     static unsigned long MAX_NUM_STATES;        // Maximal number of considered states
-    static int MAX_PLANNING_TIME;               // Maximal algorithm runtime in [ms]
+    static float MAX_PLANNING_TIME;             // Maximal algorithm runtime in [s]
     static int TERMINATE_WHEN_PATH_IS_FOUND;    // Whether to terminate when path is found (default: false)
 };
