+++ include/configurations/RGBTConnectConfig.h
@@ -7,6 +7,6 @@ class RGBTConnectConfig
 public:
     static unsigned long MAX_NUM_ITER;          // Maximal number of algorithm iterations
     static unsigned long MAX_NUM_STATES;        // Maximal number of considered states
-    static int MAX_PLANNING_TIME;               // Maximal algorithm runtime in [ms]
+    static float MAX_PLANNING_TIME;             // Maximal algorithm runtime in [s]
     static int NUM_LAYERS;                      // Number of layers (extensions) for generating a generalized bur
 };
