+++ src/planners/PlannerInfo.cpp
@@ -12,12 +12,12 @@ PlannerInfo::~PlannerInfo()
 	routine_times.clear();
 }
 
-void PlannerInfo::addIterationTime(int time)
+void PlannerInfo::addIterationTime(float time)
 {
 	iteration_times.emplace_back(time);
 }
 
-void PlannerInfo::addStateTimes(const std::vector<int> &state_times)
+void PlannerInfo::addStateTimes(const std::vector<float> &state_times)
 {
 	for (size_t i = 0; i < state_times.size(); i++)
 		PlannerInfo::state_times.emplace_back(state_times[i]);
@@ -29,10 +29,10 @@ void PlannerInfo::addCostConvergence(const std::vector<float> &cost_convergence)
 		PlannerInfo::cost_convergence.emplace_back(cost_convergence[i]);
 }
 
-void PlannerInfo::addRoutineTime(int time, int idx)
+void PlannerInfo::addRoutineTime(float time, int idx)
 {
 	for (int i = routine_times.size(); i <= idx; i++)
-		routine_times.emplace_back(std::vector<int>());
+		routine_times.emplace_back(std::vector<float>());
 	routine_times[idx].emplace_back(time);
 }
